Refactor Assets page for better maintainability and componentization.
- Extract `FileCard` component.
- Create `useAssetColumns` and `useCreateLinkLogic` hooks.
- Move utility functions to `utils` files.
- Memoize filters in `Assets.tsx`.
- Centralize types.
