+++ tavern/internal/www/src/pages/assets/Assets.tsx
@@ -1,4 +1,4 @@
-import { useState } from "react";
+import { useState, useMemo } from "react";
 import { useAssets } from "./useAssets";
 import AssetsHeader from "./components/AssetsHeader";
 import AssetsTable from "./components/AssetsTable";
@@ -11,11 +11,15 @@ import { useFilters } from "../../context/FilterContext";
 export const Assets = () => {
     const rowLimit = 10;
     const { filters } = useFilters();
-    const where: any = filters.assetName ? { nameContains: filters.assetName } : {};
 
-    if (filters.creatorId) {
-        where.hasCreatorWith = [{ id: filters.creatorId }];
-    }
+    const where = useMemo(() => {
+        const whereClause: any = filters.assetName ? { nameContains: filters.assetName } : {};
+
+        if (filters.creatorId) {
+            whereClause.hasCreatorWith = [{ id: filters.creatorId }];
+        }
+        return whereClause;
+    }, [filters.assetName, filters.creatorId]);
 
     const { assets, loading, error, totalCount, pageInfo, refetch, updateAssets, page, setPage } = useAssets(rowLimit, where);
 
@@ -26,10 +30,6 @@ export const Assets = () => {
     const handleCreateLink = (assetId: string, assetName: string) => {
         setSelectedAsset({ id: assetId, name: assetName });
         setCreateLinkModalOpen(true);
-    };
-
-    const handleRefresh = () => {
-        refetch();
     };
 
     return (
@@ -62,7 +62,7 @@ export const Assets = () => {
                     setOpen={setCreateLinkModalOpen}
                     assetId={selectedAsset.id}
                     assetName={selectedAsset.name}
-                    onSuccess={handleRefresh}
+                    onSuccess={refetch}
                 />
             )}
 
@@ -70,7 +70,7 @@ export const Assets = () => {
                 <UploadAssetModal
                     isOpen={uploadAssetModalOpen}
                     setOpen={setUploadAssetModalOpen}
-                    onUploadSuccess={handleRefresh}
+                    onUploadSuccess={refetch}
                 />
             )}
         </>
