+++ tavern/internal/www/src/pages/assets/components/CreateLinkModal.tsx
@@ -1,13 +1,10 @@
-import { FC, useState } from "react";
+import { FC } from "react";
 import Modal from "../../../components/tavern-base-ui/Modal";
 import Button from "../../../components/tavern-base-ui/button/Button";
 import AlertError from "../../../components/tavern-base-ui/AlertError";
-import { useCreateLink } from "../useAssets";
-import { format, add } from "date-fns";
 import { Clipboard } from "lucide-react";
 import { Checkbox, Tabs, TabList, TabPanels, Tab, TabPanel } from "@chakra-ui/react";
-import * as yup from "yup";
-import { useFormik } from "formik";
+import { useCreateLinkLogic } from "../hooks/useCreateLinkLogic";
 
 type CreateLinkModalProps = {
     isOpen: boolean;
@@ -17,87 +14,8 @@ type CreateLinkModalProps = {
     onSuccess?: () => void;
 };
 
-const generateRandomString = (length: number) => {
-    const chars = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
-    let result = '';
-    const randomValues = new Uint32Array(length);
-    crypto.getRandomValues(randomValues);
-    for (let i = 0; i < length; i++) {
-        result += chars[randomValues[i] % chars.length];
-    }
-    return result;
-};
-
 const CreateLinkModal: FC<CreateLinkModalProps> = ({ isOpen, setOpen, assetId, assetName, onSuccess }) => {
-    const { createLink, loading } = useCreateLink();
-    const [createdLink, setCreatedLink] = useState<string | null>(null);
-    const [error, setError] = useState<string | null>(null);
-
-    const formik = useFormik({
-        initialValues: {
-            downloadLimit: 1,
-            hasDownloadLimit: false,
-            expiryMode: 0, // 0: 10m, 1: 1h, 2: Custom
-            expiresAt: format(add(new Date(), { days: 1 }), "yyyy-MM-dd'T'HH:mm"),
-            path: generateRandomString(12),
-        },
-        validationSchema: yup.object({
-            path: yup.string().required("Path is required"),
-            downloadLimit: yup.number().when("hasDownloadLimit", {
-                is: true,
-                then: (schema) => schema.min(1, "Download limit must be at least 1").required("Download limit is required"),
-            }),
-            expiresAt: yup.string().when("expiryMode", {
-                is: 2,
-                then: (schema) => schema.required("Expiry date is required")
-                   .test("is-future", "Expiry date must be in the future", (value) => {
-                       if (!value) return false;
-                       return new Date(value) > new Date();
-                   }),
-            }),
-        }),
-        onSubmit: async (values) => {
-            setError(null);
-            let finalExpiresAt = new Date();
-            if (values.expiryMode === 0) {
-                finalExpiresAt = add(new Date(), { minutes: 10 });
-            } else if (values.expiryMode === 1) {
-                finalExpiresAt = add(new Date(), { hours: 1 });
-            } else {
-                finalExpiresAt = new Date(values.expiresAt);
-            }
-
-            try {
-                const { data } = await createLink({
-                    variables: {
-                        input: {
-                            assetID: assetId,
-                            downloadLimit: values.hasDownloadLimit ? Number(values.downloadLimit) : null,
-                            expiresAt: finalExpiresAt.toISOString(),
-                            path: values.path,
-                        },
-                    },
-                });
-
-                if (data?.createLink?.path) {
-                    const link = `${window.location.origin}/cdn/${data.createLink.path}`;
-                    setCreatedLink(link);
-                    if (onSuccess) onSuccess();
-                } else {
-                    throw new Error("Failed to create link: no path returned");
-                }
-            } catch (err: any) {
-                console.error(err);
-                setError(err.message || "An unknown error occurred");
-            }
-        },
-    });
-
-    const handleCopy = () => {
-        if (createdLink) {
-            navigator.clipboard.writeText(createdLink);
-        }
-    };
+    const { formik, createdLink, error, loading, handleClose, handleCopyAndClose } = useCreateLinkLogic({ assetId, onSuccess, setOpen });
 
     return (
         <Modal setOpen={setOpen} isOpen={isOpen} size="md">
@@ -123,21 +41,14 @@ const CreateLinkModal: FC<CreateLinkModalProps> = ({ isOpen, setOpen, assetId, a
                         </div>
                         <div className="flex justify-end gap-2 mt-2">
                             <Button
-                                onClick={() => {
-                                    setCreatedLink(null);
-                                    setOpen(false);
-                                }}
+                                onClick={handleClose}
                                 buttonVariant="outline"
                                 buttonStyle={{ color: "gray", size: "md" }}
                             >
                                 Close
                             </Button>
                             <Button
-                                onClick={() => {
-                                    handleCopy();
-                                    setCreatedLink(null);
-                                    setOpen(false);
-                                }}
+                                onClick={handleCopyAndClose}
                                 buttonVariant="solid"
                                 buttonStyle={{ color: "purple", size: "md" }}
                                 leftIcon={<Clipboard className="h-5 w-5" />}
