+++ tavern/internal/www/src/pages/assets/utils.ts
@@ -1,0 +1,97 @@
+import { LinkEdge } from "../../utils/interfacesQuery";
+
+export const truncateAssetName = (name: string, maxLength: number = 25): string => {
+    if (name.length <= maxLength) return name;
+
+    // Check for path structure (forward or backward slashes)
+    const hasPath = name.includes('/') || name.includes('\\');
+
+    if (hasPath) {
+        // Handle path truncation: prioritize keeping the filename
+        const separator = name.includes('/') ? '/' : '\\';
+        const parts = name.split(separator);
+        const fileName = parts.pop() || "";
+
+        // If filename itself is too long, truncate it
+        if (fileName.length > maxLength) {
+            return fileName.substring(0, maxLength - 3) + "...";
+        }
+
+        // Try to add parent directories until limit is reached
+        let result = fileName;
+        // Start from end of parts (deepest folder)
+        for (let i = parts.length - 1; i >= 0; i--) {
+            const part = parts[i];
+            const potential = part + separator + result;
+            // +3 for "..." prefix
+            if (potential.length + 3 <= maxLength) {
+                result = potential;
+            } else {
+                return "..." + separator + result;
+            }
+        }
+        // Should not reach here if length check passed, but fallback
+        return "..." + separator + result;
+    }
+
+    // Standard string truncation
+    return name.substring(0, maxLength - 3) + "...";
+};
+
+export const sortLinks = (links: LinkEdge[]) => {
+    return [...links].sort((a: LinkEdge, b: LinkEdge) => {
+        const now = new Date();
+        const aExpired = new Date(a.node.expiresAt) < now;
+        const bExpired = new Date(b.node.expiresAt) < now;
+
+        // 1. Unexpired first
+        if (aExpired !== bExpired) {
+            return aExpired ? 1 : -1;
+        }
+
+        // 2. Unlimited downloads first
+        const aUnlimited = a.node.downloadLimit === null;
+        const bUnlimited = b.node.downloadLimit === null;
+        if (aUnlimited !== bUnlimited) {
+            return aUnlimited ? -1 : 1;
+        }
+
+        // 3. Most remaining downloads first
+        if (!aUnlimited && !bUnlimited) {
+            const aRemaining = (a.node.downloadLimit || 0) - a.node.downloads;
+            const bRemaining = (b.node.downloadLimit || 0) - b.node.downloads;
+            return bRemaining - aRemaining;
+        }
+
+        return 0;
+    });
+};
+
+export const uploadFile = (file: File, name: string, onProgress: (progress: number) => void): Promise<void> => {
+    return new Promise((resolve, reject) => {
+        const xhr = new XMLHttpRequest();
+        xhr.open("POST", "/cdn/upload");
+
+        const formData = new FormData();
+        formData.append("fileName", name);
+        formData.append("fileContent", file);
+
+        xhr.upload.onprogress = (event) => {
+            if (event.lengthComputable) {
+                const percentComplete = (event.loaded / event.total) * 100;
+                onProgress(percentComplete);
+            }
+        };
+
+        xhr.onload = () => {
+            if (xhr.status >= 200 && xhr.status < 300) {
+                resolve();
+            } else {
+                reject(new Error(xhr.statusText || "Upload failed"));
+            }
+        };
+
+        xhr.onerror = () => reject(new Error("Network error"));
+        xhr.send(formData);
+    });
+};
